backend/requirements.txt
flask
flask-cors
openai
python-dotenv
python-docx
psycopg2-binary
